Re-orginzed index and few updates

+++ api/categories/delete.php
@@ -5,6 +5,10 @@
 
         }
         public function delete($data){
+            if(!array_key_exists('id', $data) || $data['id']==''){
+                echo json_encode(["message" => 'Missing Required Parameters']);
+                exit;
+            }
 
             
             if($this->gateway->delete($data)){
